SQL Dorks 2019/shuriken/requirements.txt
Google-Search-API
argparse
